Clean up PeriodicReportRepository
- Return results instead of unran queries
- Rename methods to represent what their query is actually doing
- Return strict types instead of Dictionary results
